requirements.txt
# conda create -n MYCHATBOT python=3.10.12
# python==3.10.12
chainlit==1.1.402
pydantic==2.9.1
arxiv==2.1.3
google-generativeai==0.8.4
#chainlit==2.1.2
#google-genai==1.0.0

